Add sidekiq-uniq-jobs to ensure EDU transfer only runs once (#319)
Because the Rake job that queues this job is run by a cron task, it
may be run on multiple servers at the same time, resulting in
concurrent connections and writes to the remote SFTP server. With this,
the job should only be run once a minute, at most, which allows for retries.