Add HCA Submission service
Uses Savon to build the request from the wsdl and xml files, though
we still use Faraday for the actual submission so that we can add the
ssl certs and keys we need. TBD if including Savon is worth it, or if
we can 'just' use Gyoku for the transform.

This also refactors some existing middleware and errors that MVI was
using and turn them into what are hopefully more generic SOAP tools
we can reuse in the future.