Do some refactoring in preperation for addition new EDU forms
Much of the existing EDU Form submission is built around the 22-1990 form,
as that's the first one that we added. With the 1990E, 5490 and 1995 coming
up we need to make the serialization code a bit more modular. This work
pulls the 22-1990 specific code into a class that isolates the functions
the erb template calls into and lays the groundwork for switching between
form types in the same spool file.

Existing conformance tests will be broken out to reflect that they're only
applicable for the 22-1990, and tests will be added for the Form classes.

There will be additional work around the model to store which type of form
the submission is, the Year-To-Date report to include the new submission
types, the JSON-Schema to be able to validate the differing types, and the
metric collection process to differentiate counts between the form types.