Add Overcommit
Overcommit will run checks against any staged code and commit messages
before allowing the code to be committed or pushed. This example
.overcommit.yml file follows what I've tended to use on past projects.

It's opt-in by running the `overcommit --install` command, and can
always be disabled or overridden if needed.

For an example of how it works, make one of the yaml configuration
files invalid and try to commit it.