Production Release 20170207a (#761)
* Revert "Hca login form changes" (#757)

* WiP: Add review instance hash as RelayState on AuthN request (#750)

* add hash as RelayState on AuthN request

* add relay state stuff to logout

* review instances run as Rails.env.production

* env name confirmed

* F/all claims (#756)

* Remove claim filter (#639)

* WS for test

* Update mock mvi response (#642)

* Update mvi yml (#646)

* Update mvi mock for melvin freeman

* Add claim type (#656)

* EVSS: WIP rename project (#686)

* EVSS: Begin file rename project

* EVSS: WIP big rename

* Rubocop

* Fix tests

* Add back routes for disability_claims

* Add back the uploaders and workers

* Add back constant for DISABILITY_CLAIMS

Refs department-of-veterans-affairs/sunsets-team#287

* Update f/all-claims from master (#693)

* Do some refactoring in preperation for addition new EDU forms

Much of the existing EDU Form submission is built around the 22-1990 form,
as that's the first one that we added. With the 1990E, 5490 and 1995 coming
up we need to make the serialization code a bit more modular. This work
pulls the 22-1990 specific code into a class that isolates the functions
the erb template calls into and lays the groundwork for switching between
form types in the same spool file.

Existing conformance tests will be broken out to reflect that they're only
applicable for the 22-1990, and tests will be added for the Form classes.

There will be additional work around the model to store which type of form
the submission is, the Year-To-Date report to include the new submission
types, the JSON-Schema to be able to validate the differing types, and the
metric collection process to differentiate counts between the form types.

* Make it easier to re-submit submissions if needed

* Re-add now-failing full_address tests

* Add additional specs for EDU form rendering

* Keep stats consistent with the existing format

* Simplify #group_submissions_by_region by using ActiveSupport

* Common SOAP client (#681)

* creates a base SOAP client for MVI and HCA

* update hca to use common soap client, added breakers support

* Prevent redirect_to from logging the session token value (#679)

* initial commit for removing the redirect_to.action_controller LogSubscriber

* finishing touches

* no longer need this initializer

* Hca controller (#668)

* fix routes

* use submit_form in controller

* allow empty values for enrollment system methods

* render success spec working

* controller should handle service errors

* capture error with raven

* lint

* dont need to stub return val

* validate form to json schema

* render json schema errors

* schema validation errors spec

* lint

* soap module renamed

* Update Brakeman (#692)

* Update Brakeman

From 3.4.1+, Brakeman reads the .ruby_version file. This will get
rid of the parse errors that we've been seeing in the output for
this. It also introduces a new `plain` type of output, which in
most cases I've seen makes the output readable on a normal sized
terminal.

* Unlock Brakeman

* Resolving conflicts

* Remove duplicated method (#712)

* Use upload date when present (#751)