Use #chomp instead of #rstrip to preserve blank lines
In some EDU templates, it's possible that we'll have multiple
blank lines at the end of a template file. While our templates
(and all files) are saved with newlines at the end that we do
need to strip out, any of those intentionally-blank lines prior
to the final one have to go. #chomp handles this for us, while
leaving us with a truncated header.