Remove the default Sidekiq error handler
The default Sidekiq error handler logs to the configured logger the
details of a job that has failed, There is no way to limit this other
then to remove the default handler. We inject a handler for Sidekiq
that exports exceptions to Sentry which will contain all information
needed to diagnose the issue so we should still be able to track any
failed jobs.