Fix an order-dependent test failure
Some specs stub out `YAML.load_file`, which I18n uses to load its
translation files. If a spec happened to run in the right order and
need to access a translation before the tested code could call it's
invocation of `YAML.load_file`, I18n would get the test data instead
of a translation file and throw an exception, failing the test.