Do some initial work to split the HCA Services
VOA, the Enrollment Service, and EE, the Enrollment & Eligiability
service live on the same systems and use the same certs, so it makes
some sense to keep them logically grouped on our side. VOA will be
used to submit claims, EE to fetch their status. Existing VCR specs
for VOA still pass, and EE will not be immediately used.