Port EVSSClaimDocumentUploader to the new Workflow framework
When the `evss_upload_workflow` Feature Flipper is enabled, EVSS
Document uploads will be sent through the Workflow code for validation,
storage, and eventual transmission. This change re-uses as many specs
as possible to ensure that the file is validated and transmitted the
same as the existing code.

Once this has been proven stable, it provides us a good basis for
dropping in document conversion, pdf stamping, virus scanning, and
whatever else we may need to do.