Initial Cut of Claim [Pension/Burial] Upload
This adds a lot of functionality to `PersistentAttachment` and `SavedClaim`, turning them into STI models that allow us to add custom data and behavior for each type of data, while retaining any shared functionality. Upon submission of a `SavedClaim`, any `PersistentAttachment` that are referenced in the payload are associated with the Claim and their `Workflow` is started, which will eventually send them to a final internal backend system for claim processing.

This also stubs out the work to generate a PDF from the JSON submission and push it throught the same Workflow as the attachments themselves.