Updates 1990 spool file template to optionally use educationProgram (#1325)
Adds support to the 1990 form for the new `educationProgram` object while
maintaining backwards compatibilty with the old structure where `school`
and `educationType` are expected to be in the root `@applicant` object.
This is accomplished using the `education_type` and `school` helpers on
the 1990 form class to select the correct available properties during
rendering.

Adds an additional `kitchen_sink_edu_prog` fixture to test 1990 spool
file generation with an `educationProgram` object. These tests and
fixtures should be completely converted over once support for the old
structure is no longer needed.