ID Card Announcement Subscriptions (#1575)
* ID Card Announcement Subscriptions

Provides endpoint to record unique emails for Veterans interested in a future announcement that VIC issues have been resolved. Super interested in name suggestions.

    POST /v0/id_card_announcement_subscription
      {
        id_card_announcement_description: {
          email: test@example.com
        }
      }

creates a record in `id_card_announcement_subscriptions` with the provided email and responds with a 202 accepted. Subsequent attempts to submit the same email will result in the same response with no additional record. The response data does not include any unique identifying information:

    { result: 'OK' }

Email format validation will occur, but should also be provided on the client side. Not sure if we should add support to https://github.com/department-of-veterans-affairs/vets-json-schema. I don't have much experience working with it, but will happily dive in to integrate there and gain that knowledge.

A simple rake task is provided to export a list of email addresses subscribed to the VIC ID card announcement list:

    rake id_card_announcement_subscriptions:export

* Fix unique index