Add cautionary note to faraday-curl in gemfile
While debugging issues related to the EVSS cross-account claims incident, I used faraday-curl to temporarily log requests to EVSS.  Cookie headers that were set within the adapter are not presented by this gem since they are not available in the middleware code path. This commit adds a note to the Gemfile to let others know that, while helpful, the curl output may not be completely accurate.

Action item from https://github.com/department-of-veterans-affairs/vets.gov-team/pull/6262