Make sure the same tags are always returned for api.auth.login benchmarks (#1626)
The way the golang prometheus client handles the conversion of statsd tags into
prometheus labels means they always need to have the same keys for a a given
metric. This is a construct that is only allowed by the golang client and the
Prometheus developers consider its use to be almost always be an anti-pattern,
so I'm not entirely sure why the statsd exported uses it.

For now this can be resolved be keeping the set of tag keys constant for given metrics.