VCR cassette reflecting a collection of multiple service episodes
An episode is inside of an `militaryServiceEpisode` element.

Sibling `militaryServiceEpisode` elements reflect a collection of these episodes.  Meaning, nothing is nested in an array, etc., to illustrate a collection.

Siblings of this `militaryServiceEpisode` type indicate a collection of them.

And if there is just one `militaryServiceEpisode`, then the collection contains only one episode.