Adding the extra definitions breaks all the swagger specs
Tried adding those extra parameter descriptions, so we could have somewhere to describe what is and is not required, in the body.

By doing so, it broke every spec in the swagger_spec.

Have not had any success finding any other way to document the content’s of the body.  Like describing with property is required, details about the given property (like validations), etc.