9706 Vet360 GET Person Integration (#1831)
This PR implements a Common::Client based service for Vet360.

lib/vet360                     - Base Vet360 client
lib/vet360/contact_information - Vet360 client for retrieving veteran contact info
app/models/vet360_redis        - Facade for caching and transforming Vet360 responses

A key component of interacting with the Vet360 service is the vet360_id, which is a
correlation id we retrieve from the MVI service. The vet360_id is how a veteran is
uniquely identified within the Vet360 service and is a pre-requisite for interacting
with the service. This id is available on the user model as `vet360_id`.

This PR also integrates Vet360 data into the user model through the `vet360_redis`
facade. This caches responses from Vet360 in redis and provides us an application
specific spot to transform Vet360 responses into data structured usefully for vets-api.