MVI error handling potential simplification
Out of a call with @saneshark where we discussed what possible
exceptional conditions we want to track vs what we want to handle
gracefully. The understanding is that there are only two cases that we
really need to wrap - when a record is not found with MVI, or when two
records are returned from MVI. Everything else represents a condition
under which we should log an exception to Sentry for tracking.

We may have definitely missed something major here, so this is just set up
to start the conversation.