Will only take effect in dev and staging, not prod
This is a key that we have set to false in production, and true the dev and staging.

This key will remain false for the next few weeks.

As this is a hot fix, this is the quickest way to get feedback in dev and staging, and not impact prod.