Only logs for the relevant bug case
We only technically need to log when the bug pops up, which is when one of the header’s value are nil.

As such, I will only gather data for this case.  At least for now, until proven we need to add more logging.

Also, there are many specs that depend on the sentry logging, and were breaking by the addition of this new sentry logging.

As such, by only logging in the dev or prod env, we are able to get the desired results.