Method now applies when user is present or not
When init_vet360_id is used for a rake task, a user will not be present, and we will be passing in individual ICNs.

When init_vet360_id is used for a current user, we get the ICN dynamically from the icn_with_aaid, and do not pass in an ICN.

This method now accounts for both use cases.

In either case, this is an asynchronous transaction on the Vet360 side, so the response is a transaction, not an actual person object.