Ensures that prometheus is not hit with an unknown exception key
We will only report to prometheus Vet360 exception keys that we know of, that are captured in our locales exceptions file.

If the key is not in that file, we do not try to ping prometheus, instead we log the key to sentry so that it can be added to the locales exceptions file.

By the by, during this check, we are hitting the singleton class where the vet360 keys have already been saved to, during the instance’s initialization.