Class dedicated to building exceptions for Vet360’s CSV file
From time to time the Vet360 folks update their error codes.  When this happens, we need to update our exceptions in keeping with these updates.

This class converts all of Vet360’s current error codes into the format that we use in our exceptions files and error reporting.

This is reusable and idempotent, for ongoing use.