Shell out to pdfinfo to validate pdfs
The Ruby PDF::Reader library has a known issue where it is unable to open pdfs
that are close to the spec, but may violate it in some minor way (see:
https://github.com/yob/pdf-reader/issues/213). This results in users sending
pdfs that they can open on their device that we reject as being invalid.

This replaces that library with the pdfinfo commandline tool based on the
poppler pdf rendering library that is more flexible with regards to pdf spec
compliance and gives us the data we need: pages counts and encryption
information.