Helps with future user/account lookup
Later on, when we want to make the association between a user and an account, we will want to search by idme_uuid.  Adding this index will makes those lookups more performant.

Also, the presence of this new unique index will supplant the need for the call of unique: true on the idme_uuid column, itself.  That and it is questionable if the string column adapter truly supports unique: true.