Creates an Account record on user login (#2230)
* Creates an Account for a user on login

If one has already been created, does nothing.

The Account#uuid is already being uniquely set as part of a before_validation callback on Account.

In an upcoming story, this will be associated with redis, to save the extra db check, once a user already has an Account record.

* Error handling ensures login flow isn’t interrupted

Though it is highly unlikely that any error could occur here, on the off chance that one does, we want to ensure that the login flow does not error out, due to the creation of this Account record.

* Temp feature flag until all account features are ready

Using this `Settings......enabled` flag approach for some of this new functionality.  That way, we can merge to master, not have to maintain a long term branch, and not officially “turn on” things on prod until we’ve got all our ducks in a row.