hack fix for dev-preview login (#2268)
This is a temporary solution to a redirect to dev.vets.gov vs dev-preview.vets.gov after the LOA1-LOA3 up-level process.

We opted for this approach to make today's code freeze deadline. We'll be working to establish a simpler URL whitelist approach - which will be documented in another ticket after this PR is accepted and merged for additional testing.

Testing included on-server dev-preview.va.gov, and dev.vets.gov.