13352 performance monitoring endpoint (#2295)
* Endpoint for posting page performance data to

* Endpoint responds with a body of what was passed to it

* Endpoint accepts multiple metrics per page

In this way, the FE can make one call to this endpoint with one page id, and multiple metrics for that page.  Vs. multiple calls to the his endpoint for each metric.

* Adds whitelisting cases to spec coverage

* Endpoint now receives params with a JSON string of data

The FE will be calling this endpoint using Navigator.sendBeacon().  This method cannot pass a nested object.  Instead, it can pass a JSON blob of data.

https://developer.mozilla.org/en-US/docs/Web/API/Navigator/sendBeacon#Parameters

* Authentication is not required for page performance analysis