Updates spec to represent 1st page of results
I believe the cassettes are off here, with regards to their titles.

If you are on the 1st page or search results, the cassette’s next_offset should be 20:

"next_offset"         => 20

So you’d think that the offset_20 cassette would have that; however, it has

"next_offset"         => 40

And the offset_0 cassette has:

"next_offset"         => 20

Np, I just adjusted the spec to pass with that setting. Just wanted to note it for future reference.