Add clarity to MVI timeouts
There was code that looked like it intended to set the open_timeout to 2 and the read_timeout to 10 for MVI,
but it didn't actually do that.  The default of 15 was used all along for both. Update the code to reflect that
reality.  I don't believe that we want to reduce the timeout because we regularly see timeout errors.