StatsD reporting for 429’s; and no sentry logging
We do not want sentry logging in these cases because with a large spike in 429 errors, and consequential logging per error, our sidekiq jobs start piling up, and bogging down our servers.

Reason being, every call to log in sentry is a background job.

Ideally, we want to report to sentry when a certain threshold of 429’s are hit, not on every single one.