Encapsulates all business logic for creating UserPreferences
One endpoint is being use to both create and update a user’s UserPreferences.  Reason being, for the update process, we are deleting all of their associated UserPreference records, and creating new ones.

So in both cases, we are creating brand new records for all of their current choices.

Also, the FE devs requested that the associated endpoint be able to accept an array of preference/choices.  This is so that the end user can make a number of preference decisions in one page, and the FE can send over just one request, in order to create all of the UserPreference records for all of the user’s decisions.

This class is responsible for:
- accepting an array of preference/choice pairings
- destroying any associated existing UserPreference records
- creating new UserPreference records based on the request
- returning the successfully creating pairings that were requested