User login no longer blocked by Vet360 503
As part of the FE login process, the FE calls the user endpoint.  Amongst other things, this calls the Vet360#get_person endpoint.

We experienced an issue where Vet360 returned a 503 for this endpoint in staging.  This blocked the user from logging in.

This rescue will now account for any such BackenndServiceException, and no longer block user login.