Fetches & serializes user's profile data for the '/v0/user' endpoint
Fetches and serializes all of the initialized user's profile data that is returned in the '/v0/user' endpoint.

If there are no external service errors, the status property is set to 200, and the outages property is set to nil.

If there *are* errors from any associated external services, the status property is set to 296, and serialized versions of the errors are added to the `outages` array.