Reduces the overall size of the contents to fit OJ gem constraints
The OJ gem we use for JSON parsing throws "NoMemoryError: Too deeply nested." errors when trying to dump an array of Virtus model objects.

This is resolved by converting the Virtus model objects to hashes, before dumping them in the OJ gem.

There is no need for the statuses to be in the Virtus model format, as we end up serializing them in the next step.

This change could have been made in the `Service.statuses_for` method; however, it feels like those class level methods in the Service file should not be concerned with formatting type issues, like converting to a hash.

The response class does feel like the correct place to make any final formatting changes to the final response that is delivered.  That is why I elected to make the change here.