Status page: Creates Vets-API external service statuses endpoint (#2905)
* Encapsulates repeated spec setup in a shared context

* Extends middleware to recognize broader swath of custom exception keys

This :raise_error middleware is used in many Faraday configurations in the code base.  In order for the middleware to leverage all of the custom error codes in the ‘config/locales/exceptions’ file, this `#service_i18n_key` must match up with a key from the locales files.

When the given error does not have a `body['code’]` in it, it resolves to nil.  This eliminates the possibility of finding a specific namespaced match in the locales file.

The only alternative to match would be adding a custom `handle_error` method at the service object level, and adding in a bunch of conditional logic there.

This refactoring gives an opportunity to provide some slightly more specific custom error keys in the locales file, and have them recognized during the middleware process.

For example, a key like `PAGERDUTY_429` would be matched on.

* Creates new v0/backend_statuses endpoint, serializer & route