Ensures that the enum of parsed statuses are consistent/transparent
We have a specific set of parsed statuses that the `GET /v0/health_care_applications/enrollment_status` endpoint can return.

If these statuses were to ever change, or be added to in the future, we want to ensure their integrity, have a single source of truth for them, etc.  For the benefits of fellow devs, ourselves in 6 months from now, etc.