Adds effective date to /v0/health_care_applications/enrollment_status response (#2955)
* Parses effective_date from E&E API response

* Adds effective_date to ../enrollment_status response

* Ensures that the enum of parsed statuses are consistent/transparent

We have a specific set of parsed statuses that the `GET /v0/health_care_applications/enrollment_status` endpoint can return.

If these statuses were to ever change, or be added to in the future, we want to ensure their integrity, have a single source of truth for them, etc.  For the benefits of fellow devs, ourselves in 6 months from now, etc.

* Adds effective_date and enum of values for parsed_status

This way our swagger docs will stipulate the precise set of potential parsed_status values, for consumers of this endpoint, as well as other devs.