Favors cached account before db call
user.account_uuid makes a .find_by db call.

We are caching a user’s account record, so using:

Account.cache_or_create_by!(user)

will first check if the account record is in the cache before making any database calls.

Under the hood this calls the Rails find_or_create_by! method.  So on the off chance there is a Redis or rails error, we follow the same rescue pattern as the rest of the class, in order to ensure that the login flow is not blocked.  Rather, the errors are reported in the meta.errors hash.