splitting session controller stats calls into 2 sets
There are a handful of stats metrics (those prefixed with STATSD_SSO)
that are used in alerts to determine failure rates for user's starting
an SSO session, while there are another set of stats (those prefixed
with STATSD_LOGIN) that are useful for indicating what types of users
are logging in and at what frequency.

The STATSD_SSO stats need to take into account every invocation of this
process, as the login process could fail at any step, to accurately
create alerts on.  Its common to see this happen twice per user
initialization because of the LOA1->LOA3 upleveling process.  Where the
STATSD_LOGIN stats would only want to report on this once per
initialization.