Urgent care subtypes #5713
This is a quick change to get posCode 17 and 20

Since we want the backend to combine the results from both pos searches this looked like a quick way to get it done.
This needs more testing around

| Urgent Care | Walk In | Results                 |
| ----------- | ------- | ----------------------- |
| yes         | yes     | Urgent Care and Walk in |
| yes         | no      | only Urgent Care        |
| no          | yes     | only Walk In            |
| no          | no      | Empty Array             |

since the client class is handeling the merge the controller just needs to keep doing what its doing.

department-of-veterans-affairs/va.gov-team#5713

trying to fix the lint errors
Moving a few more things around.
some of these linters are a little over the top