expanding upon the usefulness of the SSO success KPI metric (#3762)
* only emit login success statsd metrics once

the success login metrics are sometimes logged twice. there are a number
of users who can be up-leveled from LOA1 to LOA3 upon login, those users
are redirected back to the IDP for a second round and thus get "logged in"
twice

* updating statsd shared cookie metric to include any user with a shared sso cookie

* a more direct way to check if a shared sso cookie is set

* adding LOA and IDP tags to the SSO_SHARED_COOKIE metric

* updating the id.me LOA3 url

* updating v1 tests to use new LOA 3 url

* by default the SAML SSOe should hit the pint env to logout

* removing debugging breakpoint

* no longer need to check existing_user_identity because SSOe is passing though the authn_context

* SSOe is placing the highest LOA on the va_eauth_ial attribute

* fixing tags to correctly interpret the vars

* updating a few test cases to verify the new STATSD_SSO_SHARED_COOKIE stat

* updating v1 session controller tests to use ssoe urls

* fixing linting issues

* removing FIXME notes how that the saml url has been resolved

* seperating stats switch/case statments into individual functions to reduce complexity

* fixed linting issues in v1 sessions controller spec

* renaming the original rubysaml_settings factory so we dont break other test cases that depend on it

* adding back support to check for previous context on MHV and DSLogon contexts

* fixing tests in url_service_spec against the new LOA3 url

* splitting session controller stats calls into 2 sets

There are a handful of stats metrics (those prefixed with STATSD_SSO)
that are used in alerts to determine failure rates for user's starting
an SSO session, while there are another set of stats (those prefixed
with STATSD_LOGIN) that are useful for indicating what types of users
are logging in and at what frequency.

The STATSD_SSO stats need to take into account every invocation of this
process, as the login process could fail at any step, to accurately
create alerts on.  Its common to see this happen twice per user
initialization because of the LOA1->LOA3 upleveling process.  Where the
STATSD_LOGIN stats would only want to report on this once per
initialization.

* fixing url_service test case related to IDME_LOA3 url change

* updating STATSD metric tests in v0/sesions_controller_spec

Co-authored-by: Patrick Vinograd <patrick@adhocteam.us>