adding EVSSClaim.initial_user_uuid
when creating EVSSClaims we need to initialize the user_uuid to
reference the user's account id.  because our new where clause will
search on either the user's idme uuid or account id, the default
implementation of `values_for_create` will initialize the user_uuid to
an array, so we need to explicitly define what we want the user_uuid to
be.  an alternative approach would be to override the
`values_for_create` method when handling EVSSClaim active record
relations