Adds a centralized library for interacting with VBMS eFolder.
Service library for interacting with VBMS eFolder. This commit in
cludes a VBMS::Efolder::UploadService for uploading unstructured
documents, a base VBMS::Efolder::Service with some statsd helper
methods, and specs for each. The goal of this library is to help
centralize workflows around VBMS eFolder and is largely a wrapper
around the connect_vbms gem.