Remove dependabot
We've been trying dependabot for a few months. As of the latest round
of dependabot PRs, all of them include `Gemfile.lock` files that removed
the commercial Sidekiq gems. We can't merge these, so in this case it took
reviewing/closing/creating 15 new PRs which created a fair bit of work.

I checked our configuration on GitHub and we have enterprise.contribsys.com
configured, which must be why it *used* to work. No clear reason why it
no longer works.

There's no clear way to disable the configuration without removing it
altogether based on all the docs I've been able to find. The second best
option is removing it.