Add module scope to VAOS MessagesSerializer (#4105)
There are two `MessagesSerializer` classes in the application. During work
on CircleCi, which led to some changes in how RSpec loading was handled,
two specs were reliably breaking under certain conditions. In a nutshell:

Things that work:
```
$ bundle exec rspec ./modules/vaos/spec/request/messages_request_spec.rb:83
$ bundle exec rspec --pattern modules/*/spec/**/*_spec.rb --format progress ./modules/vaos/spec/request/messages_request_spec.rb:83
```

Things that fail — including the spec dir first:
```
$ bundle exec rspec --pattern spec/**/*_spec.rb --format progress ./modules/vaos/spec/request/messages_request_spec.rb:83
```
and the failure is
```
{\"errors\":[{\"title\":\"Internal server error\",\"detail\":\"Internal server error\",\"code\":\"500\",\"status\":\"500\",\"meta\":{\"exception\":\"undefined method `read_attribute_for_serialization' for #<OpenStruct:0x00007f9c7f1d2db0>\",\"backtrace\":[\"/Users/johnpaul/work/oddball/vets-api/.bundle/gems/active_model_serializers-0.10.4
```

The issue was in the order of loading since the unscope `MessagesSerializer` could
either by global scope or VAOS scope one depending on which loaded first. It was
straightforward to make it explicit to avoid future nasty surprises