Ensure sidekiq-admin is only started manually
Currently, foreman starts all jobs in the Procfile and either the
web or job process is set to zero.

`foreman start -m all=1,web=0` and `foreman start -m all=1,job=0`

The procfile-style invocation was moved to the sidekiq-admin.ru file
as a comment to remind any potential user about how to invoke it.