adding a new ssoe_ebenefits_links feature flag (#4190)
eBenefits is currently working on a change that will allow users with an
SSOe session to stay authenticated when navigating over. However this
will only work if the user is proxied via the eauth.va.gov domain.  This
flag will be used within the vets-website to determine the URL format
that should be used when linking a user to eBenefits.