Ignore CVE-2020-8161 until `rack` gem can be upgraded
We cannot upgrade the `rack` gem to handle this CVE because of a version
mismatch with the `sidekiq` gem ~v5.0. We cannot apply the patch without
forking the `rack` gem.

The Rack::Directory attack outlined in CVE-2020-8161 does not seem to be
applicable to our app. We do not use the `Rack::Directory` module in our
code and it is only used in a development gem dependency (`live-reload`)
which also seems irrelevant to production deployment.