Remove PUMA_PRELOAD environment var guard clause
During the work to scale puma, the PUMA_PRELOAD environment variable
was added to make it easier to toggle application preloading in the
staging and production environments. With months of successful running
with this variable being TRUE in staging and production it is long past
time to clean up this env var here and in the devops repo.