define HcaSubmissionFailureMailer (lowercase)
constantizing the mailers in the TransactionalEmailMailer for eagerloading outputs Hca, but HCASubmissionFailureMailer defines HCA (uppercase/inflection style). Autoloading failes because it looks for HcaSubmissionFailureMailer (constantized).

It doesn't make sense to include the HCA inflection is this PR, it made more sense to redefine Hca for HCASubmissionFailureMailer. Inflections will be broken out into another PR.