Renaming PPMS::Service to PPMS::Specialty
Following the JSONAPI spec for including additional data in the request
PPMS::Provider has_many PPMS::Specialty, to make the response smaller
the relationship in the objects response only contains its ID and type
after the 'data' section there is an 'included' section with all of the
relations. they are only listed once in the included section.

Linting