Add backend review group to vaos
When we made the VAOS team the owners of their module, we made it
harder for the backend review team to make updates to that module as we
always have (updating rubocop violations, changing an inflection, etc)
with our previous permissions.

This change restores the permission the backend team has always had to
the modules code.