8721 ccp v1 (#4574)
* V1 facilities CCP Endpoint

Created a new endpoint for the PPMS data that uses the fast_jsonapi,
serializer, I'm not changing the PPMS client yet to support both
object classes. all attributes internally are snake cased.

Implemented the following endpoints
 - `/v1/facilities/ccp`
 - `/v1/facilities/ccp/ID`
 - `/v1/facilities/ccp/specialties `

promoting services to an actual class

* PPMS::Specialty

Following the JSONAPI spec for including additional data in the request
PPMS::Provider has_many PPMS::Specialty, to make the response smaller
the relationship in the object's response only contains its ID and type
after the 'data' section there is an 'included' section with all of the
relations. they are only listed once in the included section.