Urgent Care Mashup (#4645)
* A new mashup endpoint to get urgent care locations

This grabs data from PPMS and Lighthouse and combines them into a single
response. at this time this doesn't appear to support pagination because
the PPMS pagination solution is a little to hacky

This is also an odd solution since `fast_jsonapi` only supports one
serializer class per response. A workaround (unpleasant hack) is to
create a dummy object that has relationships to the actual responses
and to `include` them in the response.

* Removing the code duplication

* Moving the logic for the combined api calls into the ProviderFacility model