Adding security for form data and Redis
This adds encryption/decryption for the form data as well as some serialization to make it easier to manage the encrypted data.

This logic is intended to be applied before a form is put into Redis (for Sidekiq) and then again by the Sidekiq worker when processing a form before delivering it.