Manually "olivebranch" the metadata field for Save-In-Progress data.
The FE is going to switch to only sometimes sending the inflection header
for save-in-progress routes --that is, the FE will sometimes opt-out of
olivebranch when using save-in-progress routes.

Why?
form_data sometimes has fancy keys that do not survive olivebranch's
camelCase -> snake_case -> camelCase
(or snake_case -> camelCase -> snake_case)

example:

irb(main):001:0> "Hello there, Sam-I-Am!".underscore.camelize(:lower)
=> "hello there, samIAm!"

(Sometimes sentences with spaces and punctuation end up being keys in the
json (checkboxes).)

Because the BE does not operate on form_data --it acts as a passive store-- it's
(hopefully) safe to cut out olivebranch for SIP.

From the FE, it won't look like a change:

we're going from

FE sends camelCase
     \
      \ tranformation!
       \
        `--> Database stores snake_case
            /
           / transformation!
          /
camel <--'

to:

FE sends camelCase
     \
      \ no tranformation
       \
        `--> Database stores camelCase
            /
           / no transformation
          /
camel <--'

The Catch!:

Both the FE and BE operate on the metadata field, and both are
expecting it to be in their own case.

Therefore, even when the inflection header is not sent,
the metadata field must be transformed as if olivebranch operated on
it.