Improve spec performance (#5592)
Removed integration tests related to calling the  endpoint w/ no parameters.  This was done because those particular tests were taking a long time to run and in VOAS there is not an associated use case, so there is no reason to test them.  As adjunct work, an additional ticket will be entered to review adding validation around the associated controller endpoint to only permit calls when appropriate parameters are present in the request.

Co-authored-by: brad <brad@concord.tech>