Updated scope method and extracted helper methods
The scopes endpoint used to query for all authorization servers, then
select the server whose name matched the `category` param passed to the
method. This worked okay as an MVP to get the scope endpoint up and
running, however changes made to the Okta Authorization Servers have
made this refactor necessary. We now handle the `scopes` invocation when
the category is `health` by retrieving the information on a specific
server_id, and handle all other calls the same way. For readability,
helper methods were extracted from the `scopes` method.

Signed-off-by: Braden Shipley <bradenrshipley@gmail.com>