Updated scope method and extracted helper methods (#6084)
* Updated scope method and extracted helper methods

The scopes endpoint used to query for all authorization servers, then
select the server whose name matched the `category` param passed to the
method. This worked okay as an MVP to get the scope endpoint up and
running, however changes made to the Okta Authorization Servers have
made this refactor necessary. We now handle the `scopes` invocation when
the category is `health` by retrieving the information on a specific
server_id, and handle all other calls the same way. For readability,
helper methods were extracted from the `scopes` method.

Signed-off-by: Braden Shipley <bradenrshipley@gmail.com>

* add base tests for directory service determining which method gets called

Signed-off-by: Braden Shipley <bradenrshipley@gmail.com>

* rubocop fixes

Signed-off-by: Braden Shipley <bradenrshipley@gmail.com>

* added #remove_scope_keys test and rubocop fixes

Signed-off-by: Braden Shipley <bradenrshipley@gmail.com>

* refactored tests to use VCR cassettes

Signed-off-by: Braden Shipley <bradenrshipley@gmail.com>

* removed auth server information from vcr cassettes

Signed-off-by: Braden Shipley <bradenrshipley@gmail.com>

* Refactored multiple files to address PR Review

Signed-off-by: Braden Shipley <bradenrshipley@gmail.com>