add retry scenario
Adding this retry scenario becuase it we run this rake task while the site is still up the current logic would not work. This covers both cases where attrs can be encrypted by the old key OR the new key when running the rake task. The end result will encrypt ALL records with the new key