Return true on blank bp readings data (#8631)
This is an update to the functionality introduced in #8558 - we need to check if blood pressure readings from our Lighthouse Observation request is blank, and exit the job early in those cases. The code has been updated to perform that check, and a test has been updated to ensure the expected behavior.