Update hardcoded COE form params. (#11237)
`vaHomeOwnIndicator` and `activeDutyIndicator` were both hardcoded to `false` and were never updated. This PR fixes that. `vaHomeOwnIndicator` should be true if any of the vet's prior loans were for a property that they owned. `activeDutyIndicator` should be true if the vet indicated that they are an active duty service member (ADSM).