Fix buggy rswag rake tasks (#11590)
Recent updates to the appeals_api rake tasks included expectations that
certain environment variables would be set every time the config was
loaded. These variables are not set when running unrelated rswag tasks
for the Claims API, and rake was erroring when loading the Appeals
configs in this case. These changes update the appeals api configs so
that they can be loaded successfully without the expected environment
variables.