[52258] Remove requirement that applicant be absent if unsuccessful. (#11648)
* Remove requirement that applicant be absent if unsuccessful.

Before, we assumed that if a COE's determination status was `UNABLE_TO_DETERMINE_AUTOMATICALLY` ("unsuccessful"), then an application would _not_ exist. It turns out this isn't always the case, and because we didn't handle this case, the COE status page would render a blank page. Kelli from LGY has asked us to temporarily remove the requirement that an application exist, so that all unsuccessful COEs are treated alike. If this change doesn't have any unintended consequences, then we'll make it permanent.

* Satisfy rubocop

* Satisfy rubocop.

* Update specs