Specify intent for every loan in COE form. (#11666)
The COE form was originally designed to accept a singular loan intent, regardless of how many other prior loans a vet specified in the form. We have since learned that each loan should have its own intent. This PR accompanies department-of-veterans-affairs/vets-website#23249, and sends a separate intent for each loan, to LGY.