Distinguish between LGY documents using `documentType` (#11616)
In department-of-veterans-affairs/vets-api#11335 and department-of-veterans-affairs/vets-api#11338, I started adding an ATTACHMENT tag to the description of vet-uploaded documents, to distinguish between vet-uploaded documents and notification letters. It was a hacky solution permissible at the time (and arguably still permissible). But since then, LGY has started setting the `documentType` of vet-uploaded documents to `Veteran Correspondence`. Given that, I'd like to retire our hacky solution and rely on `documentType` instead.