Ruby 3 KwArg: Form526 Convert to PDF (#11915)
Ruby 2.7 deprecated automatic casting of a hash passed as the last
argument of a method to keyword arguments. You must now use actual
keyword arguments (`foo(bar: 'baz')` instead of ` foo({bar: 'baz'})`) or cast the hash to a keyword args using the double splat (`foo(**{bar: 'baz'})`).

In preparation for upgrading to Ruby 3.0, we're implementing the new syntax.