Log error to sentry if invalid SSN or File Number sent to BGS create_person endpoint. (#12379)
It looks like many, if not all, of the errors described in department-of-veterans-affairs/va.gov-team#56333 result from a failed request to a BGS create_person (vnp_person_create, to be exact) endpoint containing the following error message: "File Number and Social Security number are different". I examined a small sample of these errors and saw that in each case a file number was present, but a series of eight asterisks were displayed in place of the ssn. I am unsure whether this means the SSN was present, or if the SSN could have been blank, or if the SSN was eight digits instead of nine. If the latter, then we may have to require veterans to have an SSN before filling out the 686c form (if that validation/check isn't done already). If the former, then we will have to figure out what to do about BGS's validation.

To help us answer this question, this PR notifies Sentry if the SSN (or file number) is absent or if the SSN is not nine digits.

As an additional note, if a file number is supposed to be the same as an SSN, then it is problematic that BGS is returning the file_number in their error response, and it is problematic that we are logging it. We will have to follow up on that. Here is a ticket for that: department-of-veterans-affairs/va.gov-team#56528.