API-24186 New max dimensions for Benefits Intake API (#12457)
* Fix benefits intake page size reporting

Previously the height and width were reversed

* API-24186 Larger max page dimensions for Benefits Intake

We've learned that pages may be up to 78 in. x 101 in., so this updates
the maximum page dimensions in the benefits intake API accordingly. This
update is behind the fetaure flag,
`vba_documents_larger_page_size_limit`. It obsoletes the old feature
flag `vba_documents_skip_dimension_check`.

* Restrict vba_documents Flipper flag usage to vba_documents module

* Add specs for CentralMail::UploadError