56995 Strip dashes from valid BGS file numbers and log invalid SSNs and file numbers. (#12530)
We have been getting errors on BGS's vnp_person_create endpoint because the file number and SSN that we send over do not match. We get our file numbers from BGS and get our SSNs from MPI or another VA.gov identity provider. A mismatch may exist because MPI or BGS have the wrong SSN or file number, because BGS has an invalid file number, or because BGS has a valid file number that is incorrectly formatted.

Where MPI or BGS have the wrong file number, or where BGS has an invalid file number, all we can do is log and report the mismatches to them so they can fix the issue. Where BGS returns a valid file number that is incorrectly formatted (i.e. contains dashes and digits, instead of just digits) we can simply reformat the SSN (i.e. remove the dashes) before submitting it to BGS along with the file number. These are the aims of this PR. See the code comments for more detail.