Duplicate EVSS Supporting Evidence Attachment infrastructure for Lighthouse
When moving our document submission to Lighthouse, we can't use the SupportingEvidenceAttachment class currently referenced in the app/workers/evss/disability_compensation_form/submit_uploads.rb worker. This is because those attachments are associated with an uploader, SupportingEvidenceAttachmentUploader, which inherits from EVSSClaimDocumentUploaderBase and uses its CarrierWave logic. EVSSClaimDocumentUploaderBase is being replaced by LighthouseDocumentUploaderBase, which has Lighthouse API-specific document upload behavior.

Therefore, if we want to reference a Lighthouse-specific evidence attachment in the upload worker, we need to duplicate both SupportingEvidenceAttachment and SupportingEvidenceAttachmentUploader (the subclass of EVSSClaimDocumentUploaderBase), with LighthouseSupportingEvidenceAttachment and its associated  LighthouseSupportingEvidenceAttachmentUploader, which inherits from LighthouseDocumentUploaderBase.

Ports over the basic tests for SupportingEvidenceAttachmentUploader; SupportingEvidenceAttachment is untested directly so there are no unit tests for LighthouseSupportingEvidenceAttachment at the moment