Refactor Form 526 User Upload endpoint for Lighthouse migration
Refactors the supporting evidence attachement uploads controller and adds the Lighthouse migration feature flag so we can save the correct attachment type when a User uploads a file to supplement their Form 526 submission - LighthouseSupportingEvidenceAttachment if the feature is enabled.

Removes the problematic FormAttachmentCreate concern used in other controllers from this controller because it tightly couples the front end caller to knowing the underlying data model for attachment records, in this case SupportingEvidenceAttachment or LighthouseSupportingEvidenceAttachment. Rather than updating the front end to branch which metadata it sends to the API, we will abandon this bad pattern entirely and have the controller expect the nested supporting_evidence_attachment param the front end currently sends regardless of the feature flag.

Preserved refactor from c84016cb55ca57c56d8fc92ec67216f8d6a874bb that extracts S3 interaction into single method, save_attachment_to_cloud! for use with the 3PI logging framework.

Refactors poorly written file class validation which was simply expecting a class name with the string UploadedFile in it. This isn't deterministic enough so I've opted to validate directly against an array of allowed classes in the VALID_FILE_CLASSES constant.

Duplicates all request specs under LighthouseSupportingEvidenceAttachment branching to ensure file type and size validatiions are preserved.

Removed spec/controllers/v0/upload_supporting_evidences_controller_spec.rb since it is tightly coupled to the FormAttachmentCreate concern and duplicates the logic we are testing in the request spec anyway.

It was also necessary to add a new serializer, LighthouseSupportingEvidenceAttachmentSerializer, to ensure the API returns attachment metadata in the same format