Update Form 0781/Form 0781a upload worker for Lighthouse
Branches PDF submission behavior in EVSS::DisabilityCompensationForm::SubmitForm0781 to submit the generated PDFs for Form 0781 and Form 0781a (if they are present) to Lighthouse if the disability_compensation_lighthouse_document_service_provider feature flag is enabled.

Elected to simply insert the Lighthouse branching at the highest level possible and preserve the existing EVSS behavior in full rather than make any refactors to clean up/avoid duplicate code. This is to avoid introducing bugs, changing the existing logging and to make it cleaner to rip out the old behavior later.

Note I did not add any Lighthouse-specific logging to the wrap_with_logging configuration in this file for now, as the upload occurs in Lighthouse::DocumentUpload and we may want to add logging there. However, we may come back and add some event logging for the upload here later in the project once we've discussed our logging strategy.