Correct Lighthouse upload implementation in SubmitForm0781 job
I had initially implemented Lighthouse behavior in this job in a similar manner to SubmitUploads: that is, expecting to retreive a file from S3 via a file name. However, this job is generating a stamped pdf on the fly and uploading that, so we need to upload the file body directly.

Note I do not yet have automated tests for this behavior, given the nature of the objects invovled I struggled to find an automated test strategy that worked; I was spinning my wheels so opting to address it again at a later date.