Update Form 526 file upload jobs to flip on user
Because Form526 submissions are tied to a specific user, we need to toggle the disability_compensation_lighthouse_document_service_provider Flipper based on the user. As such we need to retrieve the submission user and pass it to the Flipper.enabled? check in all of the jobs that were updated to post to Lighthouse.