Create Lighthouse attachment in anticipation of feature switch
Updates the controller action for saving a user-supplied Form 526 attachment to create a LighthouseSupportingEvidenceAttachment in addition to the SupportingEvidenceAttachment record it currently creates.

Form526 Document Uploads are being migrated to post to Lighthouse instead of EVSS. LighthouseSupportingEvidenceAttachment will replace SupportingEvidenceAttachment as the model for these documents.

To ensure attachments uploaded before Lighthouse support is fully enabled for the user are retreiveable on Form 526 Submissions
that occur _after_ the feature is enabled for that user, we need to create a LighthouseSupportingEvidenceAttachment record in
addition to a SupportingEvidenceAttachment for now