Remove LighthouseSupportingEvidenceAttachment support in Submit Uploads job
Refactors the SubmitUploads job to retrieve and upload documents from a SupportingEvidenceAttachment record in both the EVSS and Lighthouse cases, as we are no longer creating the unecessary LighthouseSupportingEvidenceAttachment record when a user initially uploads them.

Also cleans up some of the code flow to make it easier to follow and easier to remove the EVSS behavior when we sunset that support.